% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auto}
\alias{auto}
\title{auto}
\format{
A data frame with 392 rows and 8 variables
}
\source{
\url{http://github.com/profyliu/bsnsing}
}
\usage{
auto
}
\description{
A test data set.
}
\keyword{datasets}
