% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BreastCancer}
\alias{BreastCancer}
\title{BreastCancer}
\format{
A data frame with 699 rows and 10 variables.
}
\source{
\url{http://github.com/profyliu/bsnsing}
}
\usage{
BreastCancer
}
\description{
A test data set.
}
\keyword{datasets}
