% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{bscontrol}
\alias{bscontrol}
\title{Define Parameters for the \code{\link{bsnsing}} Fit}
\usage{
bscontrol(lambda = 1L, bin.size = 5, max.rules = Inf, nseg.numeric = 10,
  nseg.factor = 20, num2factor = 5, LP.epsilon = 1e-08, node.size = 20,
  stop.prob = 0.9, opt.solver = c("cplex", "lpSolve"),
  opt.model = c("mip", "hybrid", "lp"), bigM = 1, verbose = F)
}
\arguments{
\item{lambda}{the penalty multiplier of the L1 norm of the solution vector.}

\item{bin.size}{the minimum number of observations required in a binarization bucket.}

\item{max.rules}{the upper bound of the L1 norm of the solution vector.}

\item{nseg.numeric}{the maximum number of segments the range of a numeric variable is divided into for each inequality direction.}

\item{nseg.factor}{the maximum number of unique levels allowed in a factor variable.}

\item{num2factor}{an equality binarization rule will be created for each unique value of a numeric variable (in addition to the inequality binarization attempt), if the number of unique values of the numeric variable is less than \code{num2factor}.}

\item{LP.epsilon}{any element in the LP solution \code{w} smaller than \code{LP.epsilon} will be taken as zero.}

\item{node.size}{if the number of training cases falling into a tree node is less than \code{node.size}, the node will become a leaf and no further split will be attempted on it.}

\item{stop.prob}{if the proportion of the majority class in a tree node is greater than \code{stop.prob}, the node will become a leaf and no further split will be attempted on it.}

\item{opt.solver}{a character string in the set {'cplex', 'lpSolve'} indicating the optimization solver to be used in the program. The choice of 'cplex' requires the package \code{\link[cplexAPI]{cplexAPI}} and the choice of 'lpSolve' requires the package \code{\link[lpSolve]{lpSolve}}. The default is 'cplex'.}

\item{opt.model}{a character string in the set {'mip', 'hybrid', 'lp'} indicating the optimization model to solve in the program. The default is 'mip'. The choice of 'lp' is faster but may sacrifice the classification accuracy.}

\item{bigM}{a positive integer representing the big M value used in the MIP formulation. The default is 1.}

\item{verbose}{a logical value (TRUE or FALSE) indicating whether the solution details are to be printed on the screen.}
}
\value{
An object of class \code{\link{bscontrol}}.
}
\description{
Define Parameters for the \code{\link{bsnsing}} Fit
}
\examples{
bscontrol()  # display the default parameters
bsc <- bscontrol(stop.prob = 0.8, nseg.numeric = 10, verbose = T)
bsc
}
