% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{get_os}
\alias{get_os}
\title{Get OS type (windows, osx, linux)
This function is not exposed to bsnsing users.}
\usage{
get_os()
}
\description{
Get OS type (windows, osx, linux)
This function is not exposed to bsnsing users.
}
