% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GlaucomaMVF}
\alias{GlaucomaMVF}
\title{GlaucomaMVF}
\format{
A data frame with 170 rows and 67 variables:
}
\source{
\url{http://github.com/profyliu/bsnsing}
}
\usage{
GlaucomaMVF
}
\description{
A test dataset
}
\keyword{datasets}
