% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{import_external_rules}
\alias{import_external_rules}
\title{Import split rules from other packages}
\usage{
import_external_rules(x, y, verbose)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a binary vector.}

\item{verbose}{a logical value.}
}
\value{
a vector of character strings.
}
\description{
Run other decision tree functions available in the system and collect the node split rules for optimal selection in bsnsing.
}
