% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{iris}
\format{
A data frame with 150 rows and 5 variables.
}
\source{
\url{http://github.com/profyliu/bsnsing}
}
\usage{
iris
}
\description{
A test data set.
}
\keyword{datasets}
