% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\docType{class}
\name{mbsnsing-class}
\alias{mbsnsing-class}
\alias{mbsnsing}
\title{A class that contains multi-class classification model built by bsnsing. Can be used in summary and predict functions.}
\description{
A class that contains multi-class classification model built by bsnsing. Can be used in summary and predict functions.
}
