% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{plot.bsnsing}
\alias{plot.bsnsing}
\title{Generate latex code for plotting the bsnsing tree}
\usage{
\method{plot}{bsnsing}(
  object,
  file = "",
  class_label = c("Neg", "Pos"),
  node_color = c("red", "green"),
  verbose = F,
  ...
)
}
\arguments{
\item{object}{an object of class \code{\link{bsnsing}}.}

\item{file}{a writable connection or a character string naming the file to write to. If not supplied, the output will be written to the console.}

\item{class_label}{a character vector of two elements describing leaf node labels (for 0 and 1)}

\item{node_color}{a character vector of two elements describing node colors (for 0 and 1)}
}
\value{
NA
}
\description{
If the file argument is supplied, this function will invoke the external programs latex, dvips and ps2pdf. If these programs are not available, only the latex code will be generated. If the file argument is left empty, the latex code will be written to the console screen. The latex code utilizes the following  packages: pstricks, pst-node, pst-tree.
}
\examples{
# Suppose bs is a bsnsing object
plot(bs)
plot(bs, file = "/path/to/destination/filename.tex")
}
