% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{plot.mbsnsing}
\alias{plot.mbsnsing}
\title{Generate latex code for plotting the bsnsing tree}
\usage{
\method{plot}{mbsnsing}(object)
}
\arguments{
\item{object}{an object of class \code{\link{mbsnsing}}.}
}
\description{
Generate latex code for plotting the bsnsing tree
}
