% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.bscontrol}
\alias{print.bscontrol}
\title{Print the Object of Class \code{\link{bscontrol}}}
\usage{
\method{print}{bscontrol}(x = bscontrol(), ...)
}
\arguments{
\item{x}{an object of class \code{\link{bscontrol}}.}

\item{...}{further arguments to the print function.}
}
\description{
Print the Object of Class \code{\link{bscontrol}}
}
