% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.bsnsing}
\alias{print.bsnsing}
\title{Print the Object of Class \code{\link{bsnsing}}}
\usage{
\method{print}{bsnsing}(object, print.call = T, ...)
}
\arguments{
\item{object}{an object of class \code{\link{bsnsing}}.}
}
\description{
Print the Object of Class \code{\link{bsnsing}}
}
