% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.mbsnsing}
\alias{print.mbsnsing}
\title{Print the Object of Class \code{\link{mbsnsing}}}
\usage{
\method{print}{mbsnsing}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{mbsnsing}}.}

\item{...}{further arguments.}
}
\description{
Print the Object of Class \code{\link{mbsnsing}}
}
