% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.summary.bsnsing}
\alias{print.summary.bsnsing}
\title{Print the Summary of \code{\link{bsnsing}} Model Fits}
\usage{
\method{print}{summary.bsnsing}(object, print.call = T, ...)
}
\arguments{
\item{object}{an object of class \code{\link{summary.bsnsing}}.}
}
\description{
Print the Summary of \code{\link{bsnsing}} Model Fits
}
