% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{prt.bscontrol}
\alias{prt.bscontrol}
\title{Print the Object of Class \code{\link{bscontrol}}}
\usage{
prt.bscontrol(control = bscontrol())
}
\arguments{
\item{control}{an object of class \code{\link{bscontrol}}.}
}
\description{
Print the Object of Class \code{\link{bscontrol}}
}
