% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{prt.mbsnsing}
\alias{prt.mbsnsing}
\title{Print the Object of Class \code{\link{mbsnsing}}}
\usage{
prt.mbsnsing(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{mbsnsing}}.}

\item{...}{further arguments.}
}
\description{
Print the Object of Class \code{\link{mbsnsing}}
}
