% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{prt.summary.mbsnsing}
\alias{prt.summary.mbsnsing}
\title{Print the summary of \code{\link{mbsnsing}} model fits}
\usage{
prt.summary.mbsnsing(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{summary.mbsnsing}}.}

\item{...}{further arguments.}
}
\description{
Print the summary of \code{\link{mbsnsing}} model fits
}
