% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_bslearn}
\alias{rcpp_bslearn}
\title{C implementation of the bslearn function}
\usage{
rcpp_bslearn(
  nrows,
  ncols,
  input_y,
  input_x,
  grp,
  max_rules,
  max_time,
  node_size,
  no_same_gender_children,
  verbose
)
}
\arguments{
\item{nrows}{number of rows}

\item{ncols}{number of columns}

\item{input_y}{Integer vector}

\item{input_x}{Integer matrix}

\item{grp}{group membership indices}

\item{max_rules}{max number of rules}

\item{max_time}{max running time in seconds}

\item{node_size}{minimum node size}

\item{no_same_gender_children}{if TRUE child nodes must be of different predicted classes}

\item{verbose}{verbose level}
}
\value{
a list containing four elements
}
\description{
C implementation of the bslearn function
}
