% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{show.mbsnsing}
\alias{show.mbsnsing}
\title{Generate latex code for plotting the bsnsing tree}
\usage{
show.mbsnsing(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{mbsnsing}}.}

\item{...}{further arguments to the plot function.}
}
\description{
Generate latex code for plotting the bsnsing tree
}
