% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{summary.bsnsing}
\alias{summary.bsnsing}
\title{Summarize the bsnsing Model Fits}
\usage{
\method{summary}{bsnsing}(object = stop("no 'object' arg"), ...)
}
\arguments{
\item{object}{an object of class \code{\link{bsnsing}}.}

\item{...}{further arguments.}
}
\value{
a list of detailed information in the given \code{object}.
}
\description{
Summarize the bsnsing Model Fits
}
