% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{summary.mbsnsing}
\alias{summary.mbsnsing}
\title{Summarize mbsnsing Model Fits}
\usage{
\method{summary}{mbsnsing}(object = stop("no 'object' arg"), ...)
}
\arguments{
\item{object}{an object of class \code{\link{mbsnsing}}.}

\item{...}{further arguments.}
}
\value{
a list of detailed information in the given \code{object}.
}
\description{
Summarize mbsnsing Model Fits
}
