% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{bscontrol}
\alias{bscontrol}
\title{Define Parameters for the \code{\link{bsnsing}} Fit}
\usage{
bscontrol(
  bin.size = 5,
  nseg.numeric = 10,
  nseg.factor = 20,
  num2factor = 5,
  node.size = 10,
  stop.prob = 0.99,
  opt.solver = c("greedy", "gurobi", "lpSolve", "cplex"),
  opt.model = c("gini", "error"),
  greedy.level = 0.9,
  verbose = F
)
}
\arguments{
\item{bin.size}{the minimum number of observations required in a binarization bucket.}

\item{nseg.numeric}{the maximum number of segments the range of a numeric variable is divided into for each inequality direction.}

\item{nseg.factor}{the maximum number of unique levels allowed in a factor variable.}

\item{num2factor}{an equality binarization rule will be created for each unique value of a numeric variable (in addition to the inequality binarization attempt), if the number of unique values of the numeric variable is less than \code{num2factor}.}

\item{node.size}{if the number of training cases falling into a tree node is less than \code{node.size}, the node will become a leaf and no further split will be attempted on it.}

\item{stop.prob}{if the proportion of the majority class in a tree node is greater than \code{stop.prob}, the node will become a leaf and no further split will be attempted on it.}

\item{opt.solver}{a character string in the set {'gurobi', 'cplex', 'lpSolve', 'greedy'} indicating the optimization solver to be used in the program. The choice of 'cplex' requires the package \code{\link[cplexAPI]{cplexAPI}}, 'gurobi' requires the package \code{\link[gurobi]{gurobi}}, and 'lpSolve' requires the package \code{\link[lpSolve]{lpSolve}}. The default is 'greedy' because it is fast and does not rely on other packages.}

\item{opt.model}{a character string in the set {'gini','error'} indicating the optimization model to solve in the program. The default is 'gini'. The choice of 'error' is faster because the optimization model is smaller. The default is 'gini'.}

\item{greedy.level}{a proportion value between 0 and 1, applicable only when opt.solver is 'greedy'. In the greedy forward selection process of split rules, a candidate rule is added to the OR-clause only if the split performance (gini reduction or accuracy) after the addition multiplied by greedy.level would still be greater than the split performance before the addition. A higher value of greedy.level tend to more aggressively produce multi-variable splits.}

\item{verbose}{a logical value (TRUE or FALSE) indicating whether the solution details are to be printed on the screen.}
}
\value{
An object of class \code{\link{bscontrol}}.
}
\description{
Define Parameters for the \code{\link{bsnsing}} Fit
}
\examples{
bscontrol()  # display the default parameters
bsc <- bscontrol(stop.prob = 0.8, nseg.numeric = 10, verbose = T)
bsc
}
