% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{get_os}
\alias{get_os}
\title{Get the operating system type (windows, osx, linux).}
\usage{
get_os()
}
\value{
a character string indicating the OS type
}
\description{
This function is for internal use, to determine the file extension of the bslearn shared library. .dylib for oxs, .so for linux and .dll for windows.
}
