% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{plot.mbsnsing}
\alias{plot.mbsnsing}
\title{Generate latex code for plotting an mbsnsing tree}
\usage{
\method{plot}{mbsnsing}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{mbsnsing}}.}

\item{...}{further arguments to the plot function.}
}
\value{
No return value.
}
\description{
Users are advised to plot each bsnsing tree separately.
}
