% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.bscontrol}
\alias{print.bscontrol}
\title{Print the Object of Class \code{\link{bscontrol}}}
\usage{
\method{print}{bscontrol}(control = bscontrol())
}
\arguments{
\item{control}{an object of class \code{\link{bscontrol}}.}
}
\description{
Print the Object of Class \code{\link{bscontrol}}
}
