% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.bsnsing}
\alias{print.bsnsing}
\title{Print the Object of Class \code{\link{bsnsing}}}
\usage{
\method{print}{bsnsing}(x, print.call = T, ...)
}
\arguments{
\item{x}{an object of class \code{\link{bsnsing}}.}

\item{print.call}{print out the function called, default TRUE}

\item{...}{further arguments}
}
\description{
Print the Object of Class \code{\link{bsnsing}}
}
