% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.summary.bsnsing}
\alias{print.summary.bsnsing}
\title{Print the Summary of \code{\link{bsnsing}} Model}
\usage{
\method{print}{summary.bsnsing}(x, print.call = T, ...)
}
\arguments{
\item{x}{an object of class \code{\link{summary.bsnsing}}.}

\item{print.call}{a logical value, print out the function called if TRUE.}

\item{...}{further arguments.}
}
\description{
Print the Summary of \code{\link{bsnsing}} Model
}
