% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsnsing.R
\name{print.summary.mbsnsing}
\alias{print.summary.mbsnsing}
\title{Print the summary of \code{\link{mbsnsing}} model fits}
\usage{
\method{print}{summary.mbsnsing}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{summary.mbsnsing}}.}
}
\description{
Print the summary of \code{\link{mbsnsing}} model fits
}
